Fix for updates of clients list

+++ AWS-Rzeczy/ClientApp/src/app/home/home.component.ts
@@ -14,7 +14,8 @@ import { S3RquestBody } from './S3RquestBody';
 export class HomeComponent {
     public logs = ["nananna"];
 
-    private clients: Client[] = [new Client("log", "haselko", "Ola", 23), new Client("log2", "haselko2", "Maciej", 30), new Client("log3", "haselko3", "Kasia", 11)];
+    private startsClients: Client[] = [new Client("log", "haselko", "Ola", 23), new Client("log2", "haselko2", "Maciej", 30), new Client("log3", "haselko3", "Kasia", 11)];
+    private currentClients: Client[] = [];
     private bucketName: string;
     private key1: string;
     private key2: string;
@@ -79,22 +80,26 @@ export class HomeComponent {
     {
         this.logs = [];
         this.logs.push("-> Adding start client list...");
-        this.http.post<Message>(this.baseUrl + `api/clients/createclients`, this.clients).subscribe(
+        this.http.post<Message>(this.baseUrl + `api/clients/createclients`, this.startsClients).subscribe(
             result => {
                 this.logs.push(result.msg);
             },
             error => console.error(error));
     }
 
-    public getAllClients()
+    public getAllClients(withlogs)
     {
-        this.logs = [];
-        this.logs.push("-> Getting all clients...");
+        if (withlogs) {
+            this.logs = [];
+            this.logs.push("-> Getting all clients...");
+        }
         this.http.get<Client[]>(this.baseUrl + 'api/clients').subscribe(
             result => {
-                this.clients = result;
-                this.logs.push(`Amount of clients: ${result.length}`);
-                this.clients.forEach(client => this.logs.push(this.toString(client)));
+                this.currentClients = result;
+                if (withlogs) {
+                  this.logs.push(`Amount of clients: ${result.length}`);
+                  this.currentClients.forEach(client => this.logs.push(this.toString(client)));
+                }
             },
             error => console.error(error));
     }
@@ -107,17 +112,21 @@ export class HomeComponent {
     public updateFirstClient()
     {
         this.logs = [];
-        this.logs.push("-> Updating first client...");
-        var toUpdate = this.clients[0];
-        toUpdate.age += 1;
+        this.getAllClients(false);
+        if (this.currentClients.length > 0) {
+            this.logs.push("-> Updating first client...");
+            var toUpdate = this.currentClients[0];
+            toUpdate.age += 1;
 
-        this.http.post<Client>(this.baseUrl + `api/clients/${toUpdate.login}`, toUpdate).subscribe(
-            result => {
-                this.clients[0] = result;
-                this.logs.push(`Updated "${result.login}" age to ${result.age}`);
-                //this.logs.push(result.msg);
-            },
-            error => console.error(error));
+            this.http.post<Client>(this.baseUrl + `api/clients/${toUpdate.login}`, toUpdate).subscribe(
+                result => {
+                  this.currentClients[0] = result;
+                    this.logs.push(`Updated "${result.login}" age to ${result.age}`);
+                },
+              error => console.error(error));
+        } else {
+          this.logs.push("-> There is not enought clients...");
+        }
     }
 
     public addNewClient()
@@ -129,12 +138,11 @@ export class HomeComponent {
             password: "haselko",
             name: "Alabama",
             age: 20
-      }
+        }
         this.http.post<Message>(this.baseUrl + 'api/clients', client).subscribe(
             result => {
-            //this.clients.push(result);
-            //this.logs.push(`Added new client: ${this.toString(result)}`);
                 this.logs.push(result.msg);
+                this.getAllClients(false);
             },
             error => console.error(error));
     }
@@ -142,15 +150,20 @@ export class HomeComponent {
     public deleteLastClient()
     {
         this.logs = [];
-        this.logs.push("-> Deleting last client...");
-        var toDelete = this.clients[this.clients.length - 1];
+        this.getAllClients(false);
+        if (this.currentClients.length > 0) {
+            this.logs.push("-> Deleting last client...");
+            var toDelete = this.currentClients[this.currentClients.length - 1];
 
-        this.http.delete<Client>(this.baseUrl + `api/clients/${toDelete.login}`).subscribe(
-            result => {
-                this.clients.pop();
+            this.http.delete<Client>(this.baseUrl + `api/clients/${toDelete.login}`).subscribe(
+              result => {
                 this.logs.push(`Deleted client: ${this.toString(result)}`);
-            },
-            error => console.error(error));
+                this.getAllClients(false);
+              },
+              error => console.error(error));
+        } else {
+            this.logs.push("-> There is not enought clients...");
+        }
     }
 }
